simplify services
