+++ openhands/integrations/github/service/graphql.py
@@ -1,8 +1,3 @@
-import json
-from typing import Any
-
-import httpx
-
 from openhands.core.logger import openhands_logger as logger
 from openhands.integrations.github.queries import (
     suggested_task_issue_graphql_query,
@@ -13,37 +8,10 @@ from openhands.integrations.service_types import (
     ProviderType,
     SuggestedTask,
     TaskType,
-    UnknownException,
 )
 
 
 class GitHubGraphQLMixin(GitHubMixinBase):
-    async def execute_graphql_query(
-        self, query: str, variables: dict[str, Any]
-    ) -> dict[str, Any]:
-        try:
-            async with httpx.AsyncClient() as client:
-                github_headers = await self._get_github_headers()
-                response = await client.post(
-                    f'{self.BASE_URL}/graphql',
-                    headers=github_headers,
-                    json={'query': query, 'variables': variables},
-                )
-                response.raise_for_status()
-
-                result = response.json()
-                if 'errors' in result:
-                    raise UnknownException(
-                        f'GraphQL query error: {json.dumps(result["errors"])}'
-                    )
-
-                return dict(result)
-
-        except httpx.HTTPStatusError as e:
-            raise self.handle_http_status_error(e)
-        except httpx.HTTPError as e:
-            raise self.handle_http_error(e)
-
     async def get_suggested_tasks(self) -> list[SuggestedTask]:
         user = await self.get_user()
         login = user.login
