+++ openhands/integrations/github/service/repos.py
@@ -6,6 +6,12 @@ from openhands.integrations.service_types import OwnerType, ProviderType, Reposi
 
 
 class GitHubReposMixin(GitHubMixinBase):
+    async def get_installations(self) -> list[str]:
+        url = f'{self.BASE_URL}/user/installations'
+        response, _ = await self._make_request(url)
+        installations = response.get('installations', [])
+        return [str(i['id']) for i in installations]
+
     async def _fetch_paginated_repos(
         self, url: str, params: dict, max_repos: int, extract_key: str | None = None
     ) -> list[dict]:
